refacto(client): création du context NotificationContext qui va gérer la SnackBar
